src/visionmate/core/recognition/openai_realtime.py
"""OpenAI Realtime API client implementation.

This module provides a streaming VLM client that uses OpenAI's Realtime API
via WebSocket for low-latency multimodal interactions.
"""

from __future__ import annotations

import asyncio
import base64
import io
import logging
import time
from typing import Callable, List, Optional

import numpy as np
from openai import AsyncOpenAI
from PIL import Image

from visionmate.core.models import AudioChunk, VideoFrame
from visionmate.core.recognition.vlm_client import (
    StreamingVLMClient,
    VLMResponse,
)

logger = logging.getLogger(__name__)


class OpenAIRealtimeClient(StreamingVLMClient):
    """OpenAI Realtime API client (WebSocket streaming).

    Uses the official OpenAI Python library to connect to OpenAI's Realtime API
    for streaming multimodal interactions. Supports sending video frames and
    audio chunks in real-time and receiving streaming responses.

    Note: As of the implementation date, the OpenAI Python library's Realtime API
    support is still in development. This implementation provides the interface
    structure and will be updated when the library's Realtime API is stable.
    """

    def __init__(self, api_key: str, model: str = "gpt-4o-realtime-preview"):
        """Initialize OpenAI Realtime client.

        Args:
            api_key: OpenAI API key
            model: Model name to use (default: gpt-4o-realtime-preview)
        """
        self.api_key = api_key
        self.model = model
        self.client: Optional[AsyncOpenAI] = None
        self.response_callback: Optional[Callable[[VLMResponse], None]] = None
        self._connected = False
        self._connection_task: Optional[asyncio.Task] = None

        logger.info(f"Initialized OpenAI Realtime client with model: {model}")

    def set_model(self, model_name: str) -> None:
        """Set the VLM model to use.

        Args:
            model_name: Name of the model to use
        """
        self.model = model_name
        logger.info(f"Model set to: {model_name}")

    def get_available_models(self) -> List[str]:
        """Get list of available models.

        Returns:
            List of OpenAI Realtime model names
        """
        return [
            "gpt-4o-realtime-preview",
            "gpt-4o-realtime-preview-2024-12-17",
        ]

    async def connect(self) -> None:
        """Establish WebSocket connection to OpenAI Realtime API.

        Raises:
            ConnectionError: If connection fails
        """
        try:
            logger.info("Connecting to OpenAI Realtime API...")

            # Initialize AsyncOpenAI client
            self.client = AsyncOpenAI(api_key=self.api_key)

            # Note: The OpenAI Python library's Realtime API support is evolving.
            # This is a placeholder for the actual WebSocket connection logic
            # that will be implemented when the library's Realtime API is stable.
            #
            # Expected usage pattern (subject to change):
            # self.realtime_client = await self.client.realtime.connect(
            #     model=self.model
            # )
            #
            # For now, we mark as connected and log a warning
            self._connected = True
            logger.warning(
                "OpenAI Realtime API connection is a placeholder. "
                "Full implementation pending library support."
            )

        except Exception as e:
            logger.error(f"Failed to connect to OpenAI Realtime API: {e}")
            raise ConnectionError(f"Failed to connect: {e}") from e

    async def disconnect(self) -> None:
        """Close WebSocket connection."""
        if not self._connected:
            return

        try:
            logger.info("Disconnecting from OpenAI Realtime API...")

            # Cancel connection task if running
            if self._connection_task and not self._connection_task.done():
                self._connection_task.cancel()
                try:
                    await self._connection_task
                except asyncio.CancelledError:
                    pass

            # Close client connection
            if self.client:
                await self.client.close()

            self._connected = False
            logger.info("Disconnected from OpenAI Realtime API")

        except Exception as e:
            logger.error(f"Error during disconnect: {e}")

    async def send_frame(self, frame: VideoFrame) -> None:
        """Send video frame to VLM (streaming).

        Converts the video frame to a base64-encoded image and sends it
        to the OpenAI Realtime API.

        Args:
            frame: Video frame to send

        Raises:
            ConnectionError: If not connected
        """
        if not self._connected:
            raise ConnectionError("Not connected to OpenAI Realtime API")

        try:
            # Convert numpy array to PIL Image
            image = Image.fromarray(frame.image)

            # Convert to base64
            buffer = io.BytesIO()
            image.save(buffer, format="JPEG", quality=85)
            image_base64 = base64.b64encode(buffer.getvalue()).decode("utf-8")

            # Note: Actual sending logic will be implemented when library supports it
            # Expected pattern:
            # await self.realtime_client.send_image(
            #     image_data=image_base64,
            #     timestamp=frame.timestamp.timestamp()
            # )

            logger.debug(
                f"Frame prepared for sending: {frame.resolution.width}x{frame.resolution.height}, "
                f"source: {frame.source_id}, size: {len(image_base64)} bytes"
            )

        except Exception as e:
            logger.error(f"Error sending frame: {e}")
            raise

    async def send_audio_chunk(self, audio: AudioChunk) -> None:
        """Send audio chunk to VLM (streaming).

        Converts the audio chunk to the appropriate format and sends it
        to the OpenAI Realtime API.

        Args:
            audio: Audio chunk to send

        Raises:
            ConnectionError: If not connected
        """
        if not self._connected:
            raise ConnectionError("Not connected to OpenAI Realtime API")

        try:
            # Convert numpy array to bytes
            # OpenAI Realtime API expects PCM16 audio
            audio_bytes = (audio.data * 32767).astype(np.int16).tobytes()

            # Note: Actual sending logic will be implemented when library supports it
            # Expected pattern:
            # await self.realtime_client.send_audio(
            #     audio_data=audio_bytes,
            #     sample_rate=audio.sample_rate,
            #     channels=audio.channels,
            #     timestamp=audio.timestamp.timestamp()
            # )

            logger.debug(
                f"Audio chunk prepared for sending: {len(audio_bytes)} bytes, "
                f"sample_rate: {audio.sample_rate}, channels: {audio.channels}"
            )

        except Exception as e:
            logger.error(f"Error sending audio chunk: {e}")
            raise

    async def send_text(self, text: str) -> None:
        """Send text message to VLM (streaming).

        Args:
            text: Text message to send

        Raises:
            ConnectionError: If not connected
        """
        if not self._connected:
            raise ConnectionError("Not connected to OpenAI Realtime API")

        try:
            # Note: Actual sending logic will be implemented when library supports it
            # Expected pattern:
            # await self.realtime_client.send_text(text=text)

            logger.debug(f"Text message prepared for sending: {text[:50]}...")

        except Exception as e:
            logger.error(f"Error sending text: {e}")
            raise

    def register_response_callback(
        self,
        callback: Callable[[VLMResponse], None],
    ) -> None:
        """Register callback for receiving streaming responses.

        Args:
            callback: Function to call when response is received
        """
        self.response_callback = callback
        logger.info("Response callback registered")

    async def notify_topic_change(self) -> None:
        """Notify VLM of topic change (for Reset operation).

        This signals the VLM to reset conversation context while
        maintaining the connection.

        Raises:
            ConnectionError: If not connected
        """
        if not self._connected:
            raise ConnectionError("Not connected to OpenAI Realtime API")

        try:
            # Note: Actual reset logic will be implemented when library supports it
            # Expected pattern:
            # await self.realtime_client.reset_conversation()

            logger.info("Topic change notification sent")

        except Exception as e:
            logger.error(f"Error notifying topic change: {e}")
            raise

    async def _handle_response(self, response_data: dict) -> None:
        """Handle incoming response from OpenAI Realtime API.

        This is a placeholder for the response handling logic that will
        be implemented when the library's event system is available.

        Args:
            response_data: Response data from the API
        """
        if not self.response_callback:
            return

        try:
            # Parse response and create VLMResponse object
            vlm_response = VLMResponse(
                question=response_data.get("question"),
                direct_answer=response_data.get("direct_answer"),
                follow_up_questions=response_data.get("follow_up_questions", []),
                supplementary_info=response_data.get("supplementary_info", ""),
                confidence=response_data.get("confidence", 0.0),
                timestamp=time.time(),
                is_partial=response_data.get("is_partial", False),
            )

            # Call the registered callback
            self.response_callback(vlm_response)

        except Exception as e:
            logger.error(f"Error handling response: {e}")
