pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "visionmate"
version = "1.1.0"
description = "visionmate is a multi-modal assistant that continuously observes UVC video or screen capture and audio and streams into VLM (Vision Language Model), fusing vision inputs with spoken context to understand conversations and surface answers in real time."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.14"
dependencies = [
    "ruff",
    "ty",
    "pytest",
    "pytest-cov",
    "pytest-qt",
    "pytest-asyncio",
    "pre-commit",
    "hypothesis>=6.0.0",
    # Core capture dependencies
    "mss>=10.1.0",
    "opencv-python>=4.12.0",
    "sounddevice>=0.5.3",
    "numpy>=2.0.0,<2.3.0",
    # UI dependencies
    "PySide6>=6.10.1",
    # Web dependencies
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    # VLM dependencies
    "openai>=1.59.0",
    "openai-whisper>=20240930",
    "pillow>=11.0.0",
    # Security dependencies
    "keyring>=25.5.0",
    # Internationalization dependencies
    "babel>=2.16.0",
    # Platform-specific window detection dependencies
    "pyobjc-framework-Quartz>=10.3.1; sys_platform == 'darwin'",
    "pywin32>=308; sys_platform == 'win32'",
    "python-xlib>=0.33; sys_platform == 'linux'",
]

[tool.uv]
python-preference = "managed"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "B"]

[tool.ty.environment]
python-version = "3.14"
root = ["src"]

[tool.ty.src]
include = ["src", "tests"]

[tool.ty.terminal]
error-on-warning = false # Allow warnings for platform-specific imports

[tool.ty.rules]
# Disable unresolved-import errors for platform-specific modules
unresolved-import = "ignore"

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests/unit", "tests/integration"]
python_files = ["test_*.py", "*_test.py"]
addopts = ["-ra", "--strict-markers"]
markers = [
    "integration: integration tests that require real hardware (run in CI with Xvfb)",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
# pytest-xdist configuration
# Use -n auto to automatically detect CPU count
# Use -n <number> to specify number of workers

[dependency-groups]
dev = ["pytest-xdist>=3.8.0"]
